* update mybatis plus 3.4.3.4

+++ mybatisplus/src/main/kotlin/com/angcyo/spring/mybatis/plus/auto/IBaseAutoMybatisService.kt
@@ -30,7 +30,7 @@ interface IBaseAutoMybatisService<Table> : IBaseMybatisService<Table> {
 
     /**获取数量*/
     @LogMethodTime
-    fun autoCount(param: IAutoParam): Int {
+    fun autoCount(param: IAutoParam): Long {
         autoFill(param)
         return count(buildAutoParse().parseQuery(queryWrapper(), param))
     }
